% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{check_wercker}
\alias{check_wercker}
\title{Check for the existance of wercker apps and pipelines}
\usage{
check_wercker(repo)
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}
}
\description{
Returns a data frame with details on each of the provided repos
}
\seealso{
Other app functions: \code{\link{add_wercker}}
}
\concept{app functions}
