% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{create_wercker_org}
\alias{create_wercker_org}
\title{Create a wercker organization(s)}
\usage{
create_wercker_org(name, contact = get_wercker_email())
}
\arguments{
\item{name}{organization name(s)}

\item{contact}{organization contact email(s)}
}
\description{
Create a wercker organization(s)
}
\seealso{
Other user functions: \code{\link{get_wercker_apps}},
  \code{\link{get_wercker_email}},
  \code{\link{get_wercker_orgs}}
}
\concept{user functions}
