% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{delete_wercker_env_var}
\alias{delete_wercker_env_var}
\title{Delete environmental variable(s)}
\usage{
delete_wercker_env_var(repo, key)
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}

\item{key}{name of the environmental variable}
}
\description{
Delete environmental variable(s) from existing wercker app(s).
}
\seealso{
Other env var functions: \code{\link{add_wercker_env_var}},
  \code{\link{get_wercker_env_vars}}
}
\concept{env var functions}
