% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_wercker_apps}
\alias{get_wercker_apps}
\title{Get user's apps}
\usage{
get_wercker_apps(owner, simplify = TRUE)
}
\arguments{
\item{owner}{one or more user or organization name(s).}

\item{simplify}{return a cleaned and simplified version of the organization data frame.}
}
\description{
Returns a data frame with details on a user's organizations
}
\seealso{
Other user functions: \code{\link{create_wercker_org}},
  \code{\link{get_wercker_email}},
  \code{\link{get_wercker_orgs}}
}
\concept{user functions}
