% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{get_wercker_badge}
\alias{get_wercker_badge}
\title{Get wercker badge link}
\usage{
get_wercker_badge(repo, size = "small", type = "markdown",
  branch = "master")
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}

\item{size}{size of the badge, either small or large}

\item{type}{link format, either markdown or html}

\item{branch}{repo branch, defaults to master}
}
\description{
Generates the appropriate badge link in either HTML or Markdown format for an existing wercker app.
}
