% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_wercker_orgs}
\alias{get_wercker_orgs}
\title{Get my organization membership(s)}
\usage{
get_wercker_orgs(simplify = TRUE)
}
\arguments{
\item{simplify}{return a cleaned and simplified version of the organization data frame.}
}
\description{
Returns a data frame with details on a user's organizations
}
\seealso{
Other user functions: \code{\link{get_wercker_apps}}
}
\concept{user functions}
