% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{wercker_add_badge}
\alias{wercker_add_badge}
\title{Add wercker badge to github repo}
\usage{
wercker_add_badge(repo, badge = wercker_get_badge(repo, branch = branch),
  branch = "master", strip_existing_badge = TRUE)
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}

\item{badge}{one or more badge links, defaults to generating via \code{wercker_get_badge()}}

\item{branch}{github branch to alter}

\item{strip_existing_badge}{should any existing wercker badges be striped from the README.md?}
}
\description{
This function adds a wercker badge to a github repo's README.md.
}
