% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{wercker_add_env_var}
\alias{wercker_add_env_var}
\title{Add environmental variable}
\usage{
wercker_add_env_var(repo, key, value, protected = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}

\item{key}{name of the environmental variable}

\item{value}{value of the environmental variable}

\item{protected}{should the value be protected, proctected values cannot be retrieved.}

\item{overwrite}{should existing keys be overwritten}
}
\description{
Adds an environmental variable to an exist wercker app.
}
\seealso{
Other env var functions: \code{\link{delete_wercker_env_var}},
  \code{\link{wercker_env_vars}}
}
\concept{env var functions}
