% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{wercker_check}
\alias{wercker_check}
\title{Check for the existance of wercker apps and pipelines}
\usage{
wercker_check(repo)
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}
}
\description{
Returns a data frame with details on each of the provided repos
}
\seealso{
Other app functions: \code{\link{wercker_add}}
}
\concept{app functions}
