% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{wercker_env_vars}
\alias{wercker_env_vars}
\title{Retrieve environmental variables}
\usage{
wercker_env_vars(repo, warn = TRUE)
}
\arguments{
\item{repo}{one or more repo names in \code{owner/repo} format}

\item{warn}{enable messaging about failed retrievals (i.e. app does not exist)}
}
\description{
Retrieves all environmental variables from a wercker app.
}
\seealso{
Other env var functions: \code{\link{delete_wercker_env_var}},
  \code{\link{wercker_add_env_var}}
}
\concept{env var functions}
