% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{wercker_get_token}
\alias{wercker_get_token}
\title{Get wercker token}
\usage{
wercker_get_token()
}
\description{
\code{wercker_get_token} obtains the user's wercker authentication token.
}
\details{
This function looks for the token in the following places (in order):
\enumerate{
\item Value of \code{wercker_token} variable in \code{.wercker} environment
(this is where the package caches the token).

\item Value of \code{WERCKER_TOKEN} environmental variable.

\item Contents of \code{~/.wercker/token} file.
}
}
\examples{
\dontrun{
wercker_get_token()
}

}
\seealso{
Other authentication functions: \code{\link{wercker_set_token}},
  \code{\link{wercker_test_token}}
}
\concept{authentication functions}
