% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{wercker_local_build}
\alias{wercker_local_build}
\title{Build local repos}
\usage{
wercker_local_build(repo_dir, wercker_cli = require_wercker_cli(),
  verbose = TRUE)
}
\arguments{
\item{repo_dir}{Vector of repo directories or a single directory containing one or more repos.}

\item{wercker_cli}{Path to the wercker cli tool.}

\item{verbose}{include command output while building}
}
\description{
\code{wercker_local_build} uses the wercker cli tool to build local repos.
}
\examples{
\dontrun{
wercker_local_build('hw1/')
}

}
\concept{local repo functions}
