% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{wercker_set_token}
\alias{wercker_set_token}
\title{Set wercker token}
\usage{
wercker_set_token(token)
}
\arguments{
\item{token}{character, either the path of a file contained the token or the actual token.}
}
\description{
\code{wercker_set_token} defines the user's wercker authentication token,
this value is then accessed usin \code{wercker_get_token}
}
\examples{
\dontrun{
wercker_set_token("~/.wercker/token")
wercker_set_token("0123456789ABCDEF0123456789ABCDEF01234567")
}

}
\seealso{
Other authentication functions: \code{\link{wercker_get_token}},
  \code{\link{wercker_test_token}}
}
\concept{authentication functions}
