% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{wercker_test_token}
\alias{wercker_test_token}
\title{Test wercker token}
\usage{
wercker_test_token(token = wercker_get_token())
}
\arguments{
\item{token}{character or missing, if missing the token is obtained using \code{wercker_get_token}.}
}
\description{
\code{wercker_test_token} checks if a token is valid by attempting to authenticate with the Wercker's api.
}
\examples{
\dontrun{
wercker_test_token()
wercker_test_token("bad_token")
}

}
\seealso{
Other authentication functions: \code{\link{wercker_get_token}},
  \code{\link{wercker_set_token}}
}
\concept{authentication functions}
