% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_wercker_email}
\alias{get_wercker_email}
\title{Get wercker email}
\usage{
get_wercker_email(public = TRUE)
}
\arguments{
\item{public}{return public email or email from wercker profile}
}
\description{
Returns user's email address registered with wercker.
}
\seealso{
Other user functions: \code{\link{wercker_apps}},
  \code{\link{wercker_create_org}},
  \code{\link{wercker_orgs}}
}
\concept{user functions}
