% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wercker.R
\docType{package}
\name{wercker}
\alias{wercker}
\alias{wercker-package}
\title{wercker: A package for github based classroom and assignment management}
\description{
wercker: A package for github based classroom and assignment management
}
\section{authentication functions}{


\itemize{
\item \code{\link{get_wercker_token}} - get wercker token
\item \code{\link{set_wercker_token}} - set wercker token
\item \code{\link{test_wercker_token}} - test wercker token
}
}

